refactor chord models
