+++ src/App.tsx
@@ -6,14 +6,9 @@ import {
     PaletteColorOptions,
     Grid,
 } from "@material-ui/core";
-import { tokenize } from "./common/LyricTokenizer";
-import {
-    ChordSong,
-    ChordLine,
-    ChordBlock,
-} from "./common/ChordModel/ChordModel";
 import ChordPaper from "./components/ChordPaper";
 import { SnackbarProvider } from "notistack";
+import { NeverGonnaGiveYouUp } from "./NeverGonnaGiveYouUp";
 
 const createTheme = (): Theme => {
     const lightBlue: PaletteColorOptions = {
@@ -45,68 +40,12 @@ const createTheme = (): Theme => {
 function App() {
     const theme: Theme = createTheme();
 
-    const lyrics = [
-        "We're no strangers to love",
-        "You know the rules and so do I",
-        "A full commitment's what I'm thinking of",
-        "You wouldn't get this from any other guy",
-        "I just wanna tell you how I'm feeling",
-        "Gotta make you understand",
-        "Never gonna give you up",
-        "Never gonna let you down",
-        "Never gonna run around and desert you",
-        "Never gonna make you cry",
-        "Never gonna say goodbye",
-        "Never gonna tell a lie and hurt you",
-    ];
-
-    const chords = ["A", "Bb7", "Cm", "D/C#", "Em7", "Fmaj7", "G"];
-
-    const randomChord = (): string => {
-        return chords[Math.floor(Math.random() * chords.length)];
-    };
-
-    const chunk = (arr: string, tokenSize: number): string[] => {
-        const tokens = tokenize(arr);
-        const results: string[] = [];
-
-        for (let i = 0; i < tokens.length; i += tokenSize) {
-            const subArr = tokens.slice(i, i + tokenSize);
-            results.push(subArr.join(""));
-        }
-
-        return results;
-    };
-
-    const assembleSong = (): ChordSong => {
-        const chordLines: ChordLine[] = lyrics.map((lyricLine: string) =>
-            assembleLine(lyricLine)
-        );
-
-        return new ChordSong(chordLines);
-    };
-
-    const assembleLine = (lyrics: string): ChordLine => {
-        const lyricChunks = chunk(lyrics, 4);
-
-        const chordBlocks: ChordBlock[] = lyricChunks.map(
-            (lyricChunk: string) => {
-                return new ChordBlock({
-                    chord: randomChord(),
-                    lyric: lyricChunk,
-                });
-            }
-        );
-
-        return new ChordLine(chordBlocks);
-    };
-
     return (
         <ThemeProvider theme={theme}>
             <SnackbarProvider>
                 <Grid container justify="center">
                     <Grid item>
-                        <ChordPaper initialSong={assembleSong()} />
+                        <ChordPaper initialSong={NeverGonnaGiveYouUp()} />
                     </Grid>
                 </Grid>
             </SnackbarProvider>
