+++ src/components/Header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { Typography, Theme, Grid, Box } from "@material-ui/core";
 import { useTheme } from "@material-ui/styles";
 import EditableTypography from "./EditableTypography";
-import { ChordSong } from "../common/ChordModel/ChordModel";
+import { ChordSong } from "../common/ChordModel/ChordSong";
 
 interface HeaderProps {
     song: ChordSong;
